+++ sistema_petshop/src/main/java/com/petshop/model/Cliente.java
@@ -17,11 +17,16 @@ public class Cliente {
     // Declaração das variáveis
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
+
     private String nome;
+
     private String email;
+
     private String telefone;
+
     private String endereco;
+    
     private String fotoPath; // Caminho da imagem
 
     @OneToMany(mappedBy = "cliente", cascade = CascadeType.ALL, orphanRemoval = true)
@@ -41,11 +46,11 @@ public class Cliente {
     }
 
     // Getters and Setters
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
