+++ sistema_petshop/src/main/java/com/petshop/repository/EspecieRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.petshop.model.Especie;
 
 @Repository
-public interface EspecieRepository extends JpaRepository<Especie, Long> {
+public interface EspecieRepository extends JpaRepository<Especie, Integer> {
 }
