+++ sistema_petshop/src/main/java/com/petshop/repository/Formas_de_PagamentoRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface Formas_de_PagamentoRepository extends JpaRepository<Formas_de_Pagamento, Long> {
+public interface Formas_de_PagamentoRepository extends JpaRepository<Formas_de_Pagamento, Integer> {
 }
