+++ sistema_petshop/src/main/java/com/petshop/repository/VendedorRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.petshop.model.Vendedor;
 
 @Repository
-public interface VendedorRepository extends JpaRepository<Vendedor, Long> {
+public interface VendedorRepository extends JpaRepository<Vendedor, Integer> {
 }
